pyproject.toml
[project]
name = "llama-wrestler"
version = "0.1.0"
description = "REST endpoint tester using OpenAI-compatible APIs for LLMs"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic-ai",
    "httpx",
    "logfire",
    "pydantic-settings",
    "python-dotenv",
]

[project.scripts]
llama-wrestler = "llama_wrestler.main:main"

[dependency-groups]
dev = [
    "devtools>=0.12.2",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
